requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pillow==10.0.0
ndi-python==5.1.1.1
numpy==1.24.3
opencv-python==4.8.0.74
pyautogui==0.9.53
mss==9.0.1
python-multipart==0.0.6
gunicorn==21.2.0
jinja2==3.1.2 